[hotfix][table] Update function tests to use testing infrastructure
